refactor(datasets): decouple audio collate from VLM registry and remove label_utils
Audio ASR recipes already specify collate_fn via _target_ in YAML, so
the Qwen2_5OmniProcessor entry in VLM's COLLATE_FNS was dead code.
Remove it together with the _audio_collate_fns cross-import that
originally forced the label_utils extraction. Restore label/marker
functions inline in vlm/collate_fns.py (matching main) and have
audio/collate_fns.py import build_labels_from_template from there
directly — no circular import since VLM no longer imports audio.
